Port TestAddonCallbacks
